third_party/WebKit/LayoutTests/editing/execCommand/insert_unordered_list.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
  '<div contenteditable>',
   '^<div><br><div>',
     '<div><br></div>',
     '<div><br></div>',
    '<div></div>|',
    '</div>'
  ],
  'InsertUnorderedList',
  [
   '<div contenteditable>',
    '<div>',
      '<ul>',
        '<li>^<br></li>',
        '<li>|<br></li>',
      '</ul>',
      '<div>',
        '<div><br></div>',
        '<div></div>',
      '</div>',
    '</div>',
  '</div>'
  ]),'This tests for a hang when performing InsertUnorderedList.');
</script>
